Add separate update-command for each ad-type
For each ad-type we added a separate update-command and a validator for this command. We made this to follow the Open Closed Principle in update-command handlers and validators. These changes are similar to the changes, made in the previous commit.
For more details, see the message of the previous commit.